rename JS modules using BEM naming convention
