+++ src/js/main.js
@@ -2,5 +2,5 @@
 // REQUIRE MODULES
 
  require('./configAssetPaths.js');
- require('./loadSvg.js');
- require('./fitText.js');
+ require('./icons.js');
+ require('./media.js');
